pyproject.toml
[project]
name = "flight-ticket-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"gradio>=5.0.0",
	"openai>=1.40.0",
	"python-dotenv>=1.0.1",
]
